ensure int signs are used consistently

+++ bmc-config/src/bmc-config.h
@@ -72,8 +72,8 @@ typedef struct bmc_config_state_data
   uint8_t authentication_type_oem_proprietary;
 
   /* achu: caching to make rmcpplus priv go faster */
-  int cipher_suite_entry_count;
-  int cipher_suite_id_supported[CIPHER_SUITE_LEN];
+  uint8_t cipher_suite_entry_count;
+  uint8_t cipher_suite_id_supported[CIPHER_SUITE_LEN];
   int cipher_suite_id_supported_set;
   uint8_t cipher_suite_priv[CIPHER_SUITE_LEN];
   int cipher_suite_priv_set;
